Convert to HTTP handlers inside outside of the main package
Previously, the top-level package exported functions with the
signature `func(*util.Context) util.StatusError`, and the `main`
package was responsible for wrapping these and producing normal
HTTP handlers. However, the Google Cloud Functions execution
environment expects to be able to invoke HTTP handlers which are
exported directly from the top-level package.

We move the logic of converting these inner handlers into HTTP
handlers to the `internal/util` package, allowing us to expose
HTTP handlers from the top-level package
